+++ verl/verl/trainer/main_ppo.py
@@ -50,7 +50,7 @@ class RewardManager():
         if 'rm_scores' in data.batch.keys():
             return data.batch['rm_scores']
 
-        reward_tensor = torch.zeros_like(data.batch['responses'], dtype=torch.float32)
+        reward_tensor = torch.zeros_like(data.batch['responses'], dtype=torch.bfloat16)
 
         already_print_data_sources = {}
